requirements.txt
sqlalchemy
pandas
matplotlib
numpy
scipy

python-dotenv
requests

ydata-profiling
Ipython
ipykernel
jupyter
colorcet
plotly
bokeh
holoviews
panel
hvplot
seaborn
pingouin
scikit-learn
statsmodels
openpyxl
pydotplus
graphviz
google-cloud-translate
beautifulsoup4